BugID: 217 Some variable types changed (short must be int)

+++ Demo/CEP/TFlopCorrelator/TFC_Interface/src/DH_FIR.h
@@ -68,9 +68,9 @@ private:
   unsigned int itsBufSize;
   
   short itsFIR;
-  short itsNStations;      // #stations per buffer
-  short itsNTimes;         // #time samples per buffer
-  short itsNPol;           // #polarisations per sample
+  int itsNStations;      // #stations per buffer
+  int itsNTimes;         // #time samples per buffer
+  int itsNPol;           // #polarisations per sample
 
   RectMatrix<BufferType>* itsMatrix;
